[Common] Separate the C and C++ enum headers.
