+++ include/common/spare_enum_map.h
@@ -12,8 +12,7 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#ifndef WASMEDGE_C_API_SPARE_ENUM_MAP_H
-#define WASMEDGE_C_API_SPARE_ENUM_MAP_H
+#pragma once
 
 #include <algorithm>
 #include <array>
@@ -203,5 +202,3 @@ private:
 };
 
 } // namespace WasmEdge
-
-#endif // WASMEDGE_C_API_SPARE_ENUM_MAP_H
