+++ rpm/wasmedge.spec.in
@@ -6,7 +6,7 @@ Name:    wasmedge
 Version: %{srpm_version}
 Release: %autorelease
 Summary: High performance WebAssembly Virtual Machine
-# The entire source code is ASL 2.0 except LICENSE.spdx which is CC0
+#The entire source code is ASL 2.0 except LICENSE.spdx which is CC0
 License: ASL 2.0 and CC0
 URL:     https://github.com/%{reponame}/%{reponame}
 Source0: %{url}/releases/download/%{gittag}/%{reponame}-%{gittag}-src.tar.gz
@@ -19,7 +19,7 @@ BuildRequires: llvm-devel
 BuildRequires: ninja-build
 BuildRequires: spdlog-devel
 Requires:      llvm
-# Currently wasmedge could only be built on specific arches
+#Currently wasmedge could only be built on specific arches
 ExclusiveArch: x86_64 aarch64
 Provides: %{reponame} = %{version}-%{release}
 Provides: bundled(blake3) = 1.2.0
@@ -62,13 +62,11 @@ ln -s lib%{name}Plugin%{reponame}Process.so.%{gittag} %{buildroot}%{_libdir}/%{n
 
 %files devel
 %dir %{_includedir}/%{name}
-%{_includedir}/%{name}/dense_enum_map.h
 %{_includedir}/%{name}/enum.inc
 %{_includedir}/%{name}/enum_configure.h
 %{_includedir}/%{name}/enum_errcode.h
 %{_includedir}/%{name}/enum_types.h
 %{_includedir}/%{name}/int128.h
-%{_includedir}/%{name}/spare_enum_map.h
 %{_includedir}/%{name}/version.h
 %{_includedir}/%{name}/wasmedge.h
 %{_libdir}/lib%{name}_c.so
